Hardhat refactor 👷 (#434)
* add hardhat-deploy scaffold

* deploy: create custom contract deployer

* ts: allowJs true

* hardhat web3

* use import instead of require

* start fixing unit tests

* fix test running

* fix controller unit tests

* governor unit tests

* faucet unit tests

* Hardhatify ManagerProxy unit tests

* hardhatify merklesnapshot unit itests

* hardhatify TicketBroker tests

* hardhatify Minter tests

* hardhatify RoundsManager tests

* hardhatify PollCreator tests

* hardhatify Poll tests

* hardhatify VariableSupplyToken tests

* hardhatify ServiceRegistry tests

* refactored transcoder tests

* refactored bond tests

* refactored unbond and rebond tests

* refactored rebondFromUnbonded, withdrawStake, withdrawFees

* refactored reward, updateTranscoderWithFees, slashTranscoder

* setCurrentRoundTotalActiveStake, transcoderStatus, isActiveTranscoder, delegatorStatus, isRegisteredTranscoder, isValidUnbondingLock, getTotalBonded

* claimEarnings, claimSnapshotEarnings, pendingStake, pendingFees

* fixed event parameter

* changed accounts to signature and fixed a failing test

* bump hardhat version and add networks to config

* save proxies to deployments similar to hardhat-deploy

* hardhatify BroadcasterWithdrawalFlow integration test

* hardhatify Delegation integration test

* add genesis to deploy

* linter, solidity prettier

* husky pre-commits

* fix husky pre-commit hook

* remove prettier from pre-commit hook

* add solhint rules

* hardhatify remaining tests

* hardhatify solidity unit tests

* remove truffle migrations

* update CI to compile before running tests

* remove bn.js

* fix coverage command

* added deployment script for poll creator

* removed rpc class which used web3 to save checkpoints

* replaced web3 abiEncode with ethers pack

* Hardhat ABI Exporter plugin (#442)

* added hardhat-abi-exporter

* updated ci node version and removed truffle, ganache from dependencies

* updated node version to current lts

* fix linter for all .js and .ts files (#445)

* Code cleanup (#444)

* removed unused dependencies & scripts and updated readme

* restored docker integration test script and updated docker compose

* restored npm flag for passing arguments

+++ contracts/test/helpers/truffle/AssertBalance.sol
@@ -1,7 +1,6 @@
-pragma solidity >= 0.4.15 < 0.6.0;
+pragma solidity >=0.4.15 <0.6.0;
 
 library AssertBalance {
-
     /*
         Event: TestEvent
 
@@ -30,7 +29,11 @@ library AssertBalance {
         Returns:
             result (bool) - The result.
     */
-    function balanceEqual(address a, uint b, string memory message) public returns (bool result) {
+    function balanceEqual(
+        address a,
+        uint256 b,
+        string memory message
+    ) public returns (bool result) {
         result = (a.balance == b);
         _report(result, message);
     }
@@ -50,7 +53,11 @@ library AssertBalance {
         Returns:
             result (bool) - The result.
     */
-    function balanceNotEqual(address a, uint b, string memory message) public returns (bool result) {
+    function balanceNotEqual(
+        address a,
+        uint256 b,
+        string memory message
+    ) public returns (bool result) {
         result = (a.balance != b);
         _report(result, message);
     }
@@ -95,7 +102,7 @@ library AssertBalance {
 
     /******************************** internal ********************************/
 
-        /*
+    /*
             Function: _report
 
             Internal function for triggering <TestEvent>.
@@ -105,10 +112,7 @@ library AssertBalance {
                 message (string) - The message that is sent if the assertion fails.
         */
     function _report(bool result, string memory message) internal {
-        if(result)
-            emit TestEvent(true, "");
-        else
-            emit TestEvent(false, message);
+        if (result) emit TestEvent(true, "");
+        else emit TestEvent(false, message);
     }
-
 }
